+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## v1.5.3 (2025-10-26)
 
 -   Temporarily prohibited assigning non-builtin ball types to `player.ballType`.
+-   Defined a new type `BuiltinBallType` and exported it from `game:alias`.
 
 ## v1.5.2 (2025-10-22)